material & primitive module added
sphere moved to primitive module
material got its own file
