+++ src/primitive/mod.rs
@@ -1,0 +1,7 @@
+pub mod sphere;
+
+use ray::Ray;
+
+pub trait Primitive {
+    fn intersect(&self, ray : & mut Ray);
+}
