+++ src/ray.rs
@@ -1,7 +1,7 @@
 extern crate cgmath;
 
 use self::cgmath::{Vector3, Point3};
-use super::scene::Material;
+use material::Material;
 
 #[derive(Debug)]
 pub struct Ray {
